Switch to class-based view

+++ store/views.py
@@ -1,19 +1,22 @@
-from django.shortcuts import get_object_or_404
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
 from .models import Product, Collection
 from .serializers import CollectionSerializer, ProductSerializer
+from django.db.models.aggregates import Count
+from django.shortcuts import get_object_or_404
 from rest_framework import status
+from rest_framework.decorators import api_view
+from rest_framework.response import Response
+from rest_framework.views import APIView
+from rest_framework.mixins import ListModelMixin, CreateModelMixin
 
 
-@api_view(['GET', 'POST'])
-def product_list(request):
-    if request.method == 'GET':
+class ProductList(APIView):
+    def get(self, request):
         queryset = Product.objects.select_related('collection').all()
         serializer = ProductSerializer(
             queryset, many=True, context={'request': request})
         return Response(serializer.data)
-    elif request.method == 'POST':
+
+    def post(self, request):
         serializer = ProductSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         # serializer.validated_data
@@ -21,18 +24,21 @@ def product_list(request):
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
-@api_view(['GET', 'PUT', 'DELETE'])
-def product_detail(request, id):
-    product = get_object_or_404(Product, pk=id)
-    if request.method == 'GET':
+class ProductDetail(APIView):
+    def get(self, request, id):
+        product = get_object_or_404(Product, pk=id)
         serializer = ProductSerializer(product)
         return Response(serializer.data)
-    elif request.method == 'PUT':
+
+    def put(self, request, id):
+        product = get_object_or_404(Product, pk=id)
         serializer = ProductSerializer(product, data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data)
-    elif request.method == 'DELETE':
+
+    def delete(self, request, id):
+        product = get_object_or_404(Product, pk=id)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -40,7 +46,7 @@ def product_detail(request, id):
 @api_view(['GET', 'POST'])
 def collection_list(request):
     if request.method == 'GET':
-        queryset = Collection.objects.all()
+        queryset = Collection.objects.all().annotate(products_count=Count('products'))
         serializer = CollectionSerializer(queryset, many=True)
         return Response(serializer.data)
 
@@ -53,7 +59,8 @@ def collection_list(request):
 
 @api_view(['GET', 'PUT', 'DELETE'])
 def collection_detail(request, id):
-    collection = get_object_or_404(Collection, pk=id)
+    collection = get_object_or_404(Collection.objects.annotate(
+        products_count=Count('products')), pk=id)
     if request.method == 'GET':
         serializer = CollectionSerializer(collection)
         return Response(serializer.data)
@@ -63,5 +70,7 @@ def collection_detail(request, id):
         serializer.save()
         return Response(serializer.data)
     elif request.method == 'DELETE':
+        if collection.products.count() > 0:
+            return Response({'error': 'Non-empty collection cannot be deleted'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
         collection.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
